requirements.txt
streamlit
streamlit-extras
google-generativeai
python-dotenv
PyPDF2
Pillow
pymupdf
requests
matplotlib
seaborn
wordcloud
nltk
scikit-learn
python-docx

